// backend/src/models/enums.ts
export enum Experience {
  MoreThan10Years = ">10 Years",
  MoreThan5Years = ">5 Years",
  MoreThan2Years = ">2 Years",
  MoreThan1Year = ">1 Year",
  NoExperience = "No Experience",
}

export enum Tenure {
  Rented = "Rented",
  Paid = "Paid",
  Other = "Other",
}

export enum MaritalStatus {
  Married = "Married",
  Single = "Single",
  Widowed = "Widowed",
  Other = "Other",
}

export enum LocalityType {
  Urban = "Urban",
  Rural = "Rural",
  Other = "Other",
}

export enum Residence {
  House = "House",
  Flat = "Flat",
  Other = "Other",
}

export enum Employment {
  Employed = "Employed",
  Freelancer = "Freelancer",
  SelfEmployed = "Self-employed",
  Student = "Student",
  Unemployed = "Unemployed",
  Other = "Other",
}

export enum Weekday {
  Monday = "Monday",
  Tuesday = "Tuesday",
  Wednesday = "Wednesday",
  Thursday = "Thursday",
  Friday = "Friday",
  Saturday = "Saturday",
  Sunday = "Sunday",
}

export enum Gender {
  Male = "Male",
  Female = "Female",
  Diverse = "Diverse",
}

export enum Behaviour {
  Shy = "shy",
  Lively = "lively",
  Vigilant = "vigilant",
  Cautious = "cautious",
  Sensitive = "sensitive",
  Gentle = "gentle",
  Defensive = "defensive",
  Lazy = "lazy",
  Playful = "playful",
  Aggressive = "aggressive",
  Obedient = "obedient",
  Mischievous = "mischievous",
}